split search logic
